+++ R/ichimoku-package.R
@@ -59,7 +59,7 @@
 #'     unaffiliated with the ichimoku package.
 #'
 #'     Gao, C. (2021), \emph{ichimoku: Visualization and Tools for Ichimoku
-#'     Kinko Hyo Strategies}. R package version 1.3.0,
+#'     Kinko Hyo Strategies}. R package version 1.3.1,
 #'     \url{https://CRAN.R-project.org/package=ichimoku}.
 #'
 #' @useDynLib ichimoku, .registration = TRUE, .fixes = "ichimoku"